Static Library

+++ XuanWu/src/XuanWu/ImGui/ImGuiLayer.cpp
@@ -23,7 +23,7 @@ namespace XuanWu {
 	}
 
 	void ImGuiLayer::OnAttach()
-	{
+	{	
         // Setup Dear ImGui context
         IMGUI_CHECKVERSION();
         ImGui::CreateContext();
@@ -74,7 +74,7 @@ namespace XuanWu {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
